refactor: move issuance start logic to domain layer
- Create new start.rs module under domain/models/issuance/
- Move request/response models and business logic to domain layer
- Keep handler thin, just calling domain logic
- Restore comments in session/data.rs from main branch
