+++ src/domain/models/issuance/mod.rs
@@ -1,9 +1,14 @@
 mod error;
 mod events;
+mod start;
 mod task;
 
 pub use error::{IssuanceError, IssuanceErrorCode};
 pub use events::*;
+pub use start::{
+    start_issuance_session, StartIssuanceRequest, StartIssuanceResponse, TxCodeInfo,
+    CredentialTypeDisplay, IssuerInfo,
+};
 use parking_lot::Mutex;
 use serde::{Deserialize, Serialize};
 pub use task::{IssuanceTask, TaskResult};
